Move schema viewsets into their own file.
